Add ergonomic! global method
This can be used at the top-level to enable ergonomic mode everywhere,
or during a class definition to enable it just for that class.

This works because top-level methods are actually exposed as private
instance methods on the Object class. When you call ergonomic! on
a class object, the implementation actually runs `include Ergonomic`
with the implied receiver being the class. Not of the top-level!

Yes this quirky feature of Ruby also broke my brain.
But now I've come to appreciate it.